Remove WordPress environmental variables from global environment
PHP-FPM starts with a clean environment, negating any environmental
variables that are set outside of the service startup process. This pull
request takes the `wordpress_env` environmental variable mapping and
passes it to `php_fpm_env` so that processes under PHP-FPM can have
access to the proper environmental variable values.

This snuck by in my previous tests because the default
`wp-config-sample.php` works out of the box with the default database
name, username, and password I chose.